web3 1.2.2 & truffle-hd-wallet updates, simplify types
