+++ .eslintrc
@@ -31,12 +31,13 @@
     "env": {
         "es6": true,
         "browser": true,
+        "node": true,
         "jest": true,
         "cypress/globals": true
     },
     "settings": {
         "react": {
-            "version": "16.8"
+            "version": "16"
         }
     }
 }
