+++ client/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "react-scripts start",
     "build": "CI=false react-scripts --max_old_space_size=4096 build",
-    "test": "CI=false react-scripts test --coverage --watchAll=false",
+    "test": "react-scripts test --coverage --watchAll=false",
     "test:watch": "react-scripts test --coverage",
     "eject": "react-scripts eject",
     "coverage": "cat coverage/lcov.info | codacy-coverage --token 8801f827fe1144ffa85cd7da94f2bbf7",
@@ -17,6 +17,7 @@
     "@oceanprotocol/squid": "^0.8.0",
     "@oceanprotocol/typographies": "^0.1.0",
     "@sindresorhus/slugify": "^0.9.1",
+    "@truffle/hdwallet-provider": "^1.0.23",
     "axios": "^0.19.0",
     "bip39": "^3.0.2",
     "classnames": "^2.2.6",
@@ -44,15 +45,14 @@
     "react-router-dom": "^5.1.2",
     "react-transition-group": "^4.3.0",
     "shortid": "^2.2.15",
-    "truffle-hdwallet-provider": "1.0.14",
-    "web3": "1.2.1"
+    "truffle-hdwallet-provider": "^1.0.17",
+    "web3": "^1.2.2"
   },
   "devDependencies": {
     "@react-mock/state": "^0.1.8",
     "@testing-library/jest-dom": "^4.2.0",
     "@testing-library/react": "^9.3.0",
     "@types/classnames": "^2.2.9",
-    "@types/filesize": "^4.2.0",
     "@types/is-url": "^1.2.28",
     "@types/jest": "^24.0.20",
     "@types/react": "^16.9.11",
@@ -64,9 +64,9 @@
     "@types/react-paginate": "^6.2.1",
     "@types/react-router-dom": "^5.1.0",
     "@types/react-transition-group": "^4.2.3",
-    "@types/shortid": "0.0.29",
-    "@types/web3": "^1.0.20",
+    "@types/shortid": "^0.0.29",
     "@typescript-eslint/eslint-plugin": "^2.6.0",
+    "@typescript-eslint/parser": "^2.6.0",
     "eslint": "^5.16.0",
     "jest-mock-axios": "^3.1.2",
     "node-sass": "^4.13.0",
