--- client/src/@types/ipfs-http-client/index.d.ts
@@ -1,1 +1,0 @@
-declare module 'ipfs-http-client'
