--- client/src/@types/ipfs/index.d.ts
@@ -1,1 +1,0 @@
-declare module 'ipfs'
