+++ client/src/@types/node_modules.d.ts
@@ -3,3 +3,7 @@
 declare module 'ethereum-blockies' {
     export function toDataUrl(address: string): string
 }
+
+declare module 'ipfs'
+declare module 'ipfs-http-client'
+declare module 'react-collapsed'
