--- client/src/@types/react-collapsed/index.d.ts
@@ -1,1 +1,0 @@
-declare module 'react-collapsed'
