--- client/src/@types/truffle-hdwallet-provider'/index.d.ts
@@ -1,1 +1,0 @@
-declare module 'truffle-hdwallet-provider'
